refactor code into different components
