+++ src/components/Navbar.jsx
@@ -1,30 +1,5 @@
 import React, { useState, useEffect } from "react";
-
-const useLocalStorage = (key, initialValue) => {
-  const [storedValue, setStoredValue] = useState(() => {
-    const valueFromStorage = window.localStorage.getItem(key);
-    return valueFromStorage ? JSON.parse(valueFromStorage) : initialValue;
-  });
-
-  const setValue = value => {
-    setStoredValue(value);
-    window.localStorage.setItem(key, JSON.stringify(value));
-  };
-  return [storedValue, setValue];
-};
-
-const useDarkMode = initialValue => {
-  const [darkMode, setDarkMode] = useLocalStorage("dark-mode", initialValue);
-  useEffect(() => {
-    if (darkMode) {
-      document.querySelector("body").classList.add("dark-mode");
-    } else {
-      document.querySelector("body").classList.remove("dark-mode");
-    }
-  }, [darkMode]);
-
-  return [darkMode, setDarkMode];
-};
+import useDarkMode from "../hooks/useDarkMode";
 
 const Navbar = () => {
   const [darkMode, setDarkMode] = useDarkMode(false);
