+++ src/hooks/useLocalStorage.js
@@ -1,0 +1,16 @@
+import React, { useState } from "react";
+
+const useLocalStorage = (key, initialValue) => {
+  const [storedValue, setStoredValue] = useState(() => {
+    const valueFromStorage = window.localStorage.getItem(key);
+    return valueFromStorage ? JSON.parse(valueFromStorage) : initialValue;
+  });
+
+  const setValue = value => {
+    setStoredValue(value);
+    window.localStorage.setItem(key, JSON.stringify(value));
+  };
+  return [storedValue, setValue];
+};
+
+export default useLocalStorage;
